Reorganize spatial tools into georaster/ modules
