+++ geoutils/__init__.py
@@ -2,7 +2,6 @@
 GeoUtils is a python package of raster and vector tools.
 """
 
-from geoutils import spatial_tools  # noqa
 from geoutils import examples, georaster, geovector, projtools  # noqa
 from geoutils.georaster import Mask, Raster, SatelliteImage  # noqa
 from geoutils.geovector import Vector  # noqa
