+++ geoutils/georaster/__init__.py
@@ -1,4 +1,7 @@
 from geoutils.georaster.raster import Raster, RasterType, Mask  # noqa isort:skip
 from geoutils.georaster.satimg import SatelliteImage  # noqa
+from geoutils.georaster.array import * # noqa
+from geoutils.georaster.sampling import * # noqa
+from geoutils.georaster.multiraster import * # noqa
 
 __all__ = ["RasterType", "Raster"]
