+++ geoutils/georaster/multiraster.py
@@ -1,9 +1,4 @@
-"""Basic operations to be run on 2D image arrays
-
-Optional dependencies:
-    skimage.transform (@subdivide_array)
-
-"""
+"""Multiple rasters tools."""
 from __future__ import annotations
 
 import warnings
@@ -15,77 +10,9 @@ import rasterio.warp
 from tqdm import tqdm
 
 import geoutils as gu
-from geoutils.georaster import Raster, RasterType
+from geoutils.georaster import Raster, RasterType, get_array_and_mask
 from geoutils.georaster.raster import _default_nodata
 from geoutils.misc import resampling_method_from_str
-
-
-def get_mask(array: np.ndarray | np.ma.masked_array) -> np.ndarray:
-    """
-    Return the mask of invalid values, whether array is a ndarray with NaNs or a np.ma.masked_array.
-
-    :param array: Input array.
-
-    :returns invalid_mask: boolean array, True where array is masked or Nan.
-    """
-    mask = (array.mask | ~np.isfinite(array.data)) if isinstance(array, np.ma.masked_array) else ~np.isfinite(array)
-    return mask.squeeze()
-
-
-def get_array_and_mask(
-    array: np.ndarray | np.ma.masked_array | RasterType, check_shape: bool = True, copy: bool = True
-) -> tuple[np.ndarray, np.ndarray]:
-    """
-    Return array with masked values set to NaN and the associated mask.
-    Works whether array is a ndarray with NaNs or a np.ma.masked_array.
-
-    :param array: Input array.
-    :param check_shape: Validate that the array is either a 1D array, a 2D array or a 3D array of shape (1, rows, cols).
-    :param copy: Return a copy of 'array'. If False, a view will be attempted (and warn if not possible)
-
-    :returns array_data, invalid_mask: a tuple of ndarrays. First is array with invalid pixels converted to NaN, \
-    second is mask of invalid pixels (True if invalid).
-    """
-    if isinstance(array, gu.Raster):
-        array = array.data
-
-    if check_shape:
-        if len(array.shape) > 2 and array.shape[0] > 1:
-            raise ValueError(
-                f"Invalid array shape given: {array.shape}." "Expected 2D array or 3D array where arr.shape[0] == 1"
-            )
-
-    # If an occupied mask exists and a view was requested, trigger a warning.
-    if not copy and np.any(getattr(array, "mask", False)):
-        warnings.warn("Copying is required to respect the mask. Returning copy. Set 'copy=True' to hide this message.")
-        copy = True
-
-    # If array is of type integer and has a mask, it needs to be converted to float (to assign nans)
-    if np.any(getattr(array, "mask", False)) and np.issubdtype(array.dtype, np.integer):  # type: ignore
-        array = array.astype(np.float32)  # type: ignore
-
-    # Convert into a regular ndarray (a view or copy depending on the 'copy' argument)
-    array_data = np.array(array).squeeze() if copy else np.asarray(array).squeeze()
-
-    # Get the mask of invalid pixels and set nans if it is occupied.
-    invalid_mask = get_mask(array)
-    if np.any(invalid_mask):
-        array_data[invalid_mask] = np.nan
-
-    return array_data, invalid_mask
-
-
-def get_valid_extent(array: np.ndarray | np.ma.masked_array) -> tuple[int, ...]:
-    """
-    Return (rowmin, rowmax, colmin, colmax), the first/last row/column of array with valid pixels
-    """
-    if not array.dtype == "bool":
-        valid_mask = ~get_mask(array)
-    else:
-        valid_mask = array
-    cols_nonzero = np.where(np.count_nonzero(valid_mask, axis=0) > 0)[0]
-    rows_nonzero = np.where(np.count_nonzero(valid_mask, axis=1) > 0)[0]
-    return rows_nonzero[0], rows_nonzero[-1], cols_nonzero[0], cols_nonzero[-1]
 
 
 def load_multiple_rasters(
@@ -380,181 +307,3 @@ If several algorithms are provided, each result is returned as a separate band.
     )
 
     return merged_raster
-
-
-def _get_closest_rectangle(size: int) -> tuple[int, int]:
-    """
-    Given a 1D array size, return a rectangular shape that is closest to a cube which the size fits in.
-
-    If 'size' does not have an integer root, a rectangle is returned that is slightly larger than 'size'.
-
-    :examples:
-        >>> _get_closest_rectangle(4)  # size will be 4
-        (2, 2)
-        >>> _get_closest_rectangle(9)  # size will be 9
-        (3, 3)
-        >>> _get_closest_rectangle(3)  # size will be 4; needs padding afterward.
-        (2, 2)
-        >>> _get_closest_rectangle(55) # size will be 56; needs padding afterward.
-        (7, 8)
-        >>> _get_closest_rectangle(24)  # size will be 25; needs padding afterward
-        (5, 5)
-        >>> _get_closest_rectangle(85620)  # size will be 85849; needs padding afterward
-        (293, 293)
-        >>> _get_closest_rectangle(52011)  # size will be 52212; needs padding afterward
-        (228, 229)
-    """
-    close_cube = int(np.sqrt(size))
-
-    # If size has an integer root, return the respective cube.
-    if close_cube**2 == size:
-        return (close_cube, close_cube)
-
-    # One of these rectangles/cubes will cover all cells, so return the first that does.
-    potential_rectangles = [(close_cube, close_cube + 1), (close_cube + 1, close_cube + 1)]
-
-    for rectangle in potential_rectangles:
-        if np.prod(rectangle) >= size:
-            return rectangle
-
-    raise NotImplementedError(f"Function criteria not met for rectangle of size: {size}")
-
-
-def subdivide_array(shape: tuple[int, ...], count: int) -> np.ndarray:
-    """
-    Create indices for subdivison of an array in a number of blocks.
-
-    If 'count' is divisible by the product of 'shape', the amount of cells in each block will be equal.
-    If 'count' is not divisible, the amount of cells in each block will be very close to equal.
-
-    :param shape: The shape of a array to be subdivided.
-    :param count: The amount of subdivisions to make.
-
-    :examples:
-        >>> subdivide_array((4, 4), 4)
-        array([[0, 0, 1, 1],
-               [0, 0, 1, 1],
-               [2, 2, 3, 3],
-               [2, 2, 3, 3]])
-
-        >>> subdivide_array((6, 4), 4)
-        array([[0, 0, 1, 1],
-               [0, 0, 1, 1],
-               [0, 0, 1, 1],
-               [2, 2, 3, 3],
-               [2, 2, 3, 3],
-               [2, 2, 3, 3]])
-
-        >>> subdivide_array((5, 4), 3)
-        array([[0, 0, 0, 0],
-               [0, 0, 0, 0],
-               [1, 1, 2, 2],
-               [1, 1, 2, 2],
-               [1, 1, 2, 2]])
-
-    :raises ValueError: If the 'shape' size (`np.prod(shape)`) is smallern than 'count'
-                        If the shape is not a 2D shape.
-
-    :returns: An array of shape 'shape' with 'count' unique indices.
-    """
-    try:
-        import skimage.transform
-    except ImportError:
-        raise ImportError("Missing optional dependency, skimage.transform, required by this function.")
-
-    if count > np.prod(shape):
-        raise ValueError(f"Shape '{shape}' size ({np.prod(shape)}) is smaller than 'count' ({count}).")
-
-    if len(shape) != 2:
-        raise ValueError(f"Expected a 2D shape, got {len(shape)}D shape: {shape}")
-
-    # Generate a small grid of indices, with the same unique count as 'count'
-    rect = _get_closest_rectangle(count)
-    small_indices = np.pad(np.arange(count), np.prod(rect) - count, mode="edge")[: np.prod(rect)].reshape(rect)
-
-    # Upscale the grid to fit the output shape using nearest neighbour scaling.
-    indices = skimage.transform.resize(small_indices, shape, order=0, preserve_range=True).astype(int)
-
-    return indices.reshape(shape)
-
-
-def get_xy_rotated(raster: Raster, along_track_angle: float) -> tuple[np.ndarray, np.ndarray]:
-    """
-    Rotate x, y axes of image to get along- and cross-track distances.
-    :param raster: Raster to get x,y positions from.
-    :param along_track_angle: Angle by which to rotate axes (degrees)
-
-    :returns xxr, yyr: Arrays corresponding to along (x) and cross (y) track distances.
-    """
-
-    myang = np.deg2rad(along_track_angle)
-
-    # Get grid coordinates
-    xx, yy = raster.coords(grid=True)
-    xx -= np.min(xx)
-    yy -= np.min(yy)
-
-    # Get rotated coordinates
-
-    # For along-track
-    xxr = xx * np.cos(myang) - yy * np.sin(myang)
-    # For cross-track
-    yyr = xx * np.sin(myang) + yy * np.cos(myang)
-
-    # Re-initialize coordinate at zero
-    xxr -= np.nanmin(xxr)
-    yyr -= np.nanmin(yyr)
-
-    return xxr, yyr
-
-
-def subsample_raster(
-    array: np.ndarray | np.ma.masked_array,
-    subsample: float | int,
-    return_indices: bool = False,
-    random_state: None | np.random.RandomState | np.random.Generator | int = None,
-) -> np.ndarray:
-    """
-    Randomly subsample a 1D or 2D array by a subsampling factor, taking only non NaN/masked values.
-
-    :param subsample: If <= 1, will be considered a fraction of valid pixels to extract.
-    If > 1 will be considered the number of pixels to extract.
-    :param return_indices: If set to True, will return the extracted indices only.
-    :param random_state: Random state, or seed number to use for random calculations (for testing)
-
-    :returns: The subsampled array (1D) or the indices to extract (same shape as input array)
-    """
-    # Define state for random subsampling (to fix results during testing)
-    if random_state is None:
-        rnd = np.random.default_rng()
-    elif isinstance(random_state, (np.random.RandomState, np.random.Generator)):
-        rnd = random_state
-    else:
-        rnd = np.random.RandomState(np.random.MT19937(np.random.SeedSequence(random_state)))
-
-    # Get number of points to extract
-    if (subsample <= 1) & (subsample > 0):
-        npoints = int(subsample * np.size(array))
-    elif subsample > 1:
-        npoints = int(subsample)
-    else:
-        raise ValueError("`subsample` must be > 0")
-
-    # Remove invalid values and flatten array
-    mask = get_mask(array)  # -> need to remove .squeeze in get_mask
-    valids = np.argwhere(~mask.flatten()).squeeze()
-
-    # Checks that array and npoints are correct
-    assert np.ndim(valids) == 1, "Something is wrong with array dimension, check input data and shape"
-    if npoints > np.size(valids):
-        npoints = np.size(valids)
-
-    # Randomly extract npoints without replacement
-    indices = rnd.choice(valids, npoints, replace=False)
-    unraveled_indices = np.unravel_index(indices, array.shape)
-
-    if return_indices:
-        return unraveled_indices
-
-    else:
-        return array[unraveled_indices]
